Move distributed headers into the public API directory, in preparation for integrating with TTNN.
