+++ tt_metal/api/tt-metalium/mesh_workload.hpp
@@ -6,9 +6,7 @@
 
 #include <host_api.hpp>
 #include <mesh_device.hpp>
-
-#include "tt_metal/distributed/mesh_buffer.hpp"
-#include "tt_metal/impl/program/dispatch.hpp"
+#include <mesh_buffer.hpp>
 
 namespace tt::tt_metal::distributed {
 // The LogicalDeviceRange concept is fundamentally identical to the CoreRange concept
