+++ tt_metal/distributed/distributed.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "tt_metal/distributed/distributed.hpp"
+#include <tt-metalium/distributed.hpp>
 
 namespace tt::tt_metal::distributed {
