+++ tt_metal/distributed/mesh_buffer.cpp
@@ -3,10 +3,9 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
+#include <mesh_buffer.hpp>
 #include <overloaded.hpp>
 #include <tt_metal.hpp>
-
-#include "tt_metal/distributed/mesh_buffer.hpp"
 
 namespace tt::tt_metal::distributed {
 namespace {
